Move header detail of camapign to component
