+++ src/pages/Campaing/ShowCampaign/ShowCampaign.js
@@ -3,7 +3,8 @@ import { Query } from 'react-apollo';
 import { getCampaign, huntersCampaign, couponsByHunterInCampaign } from 'Services/graphql/queries.graphql';
 import { injectIntl } from 'react-intl';
 import classNames from 'classnames/bind';
-import { Typography, Icon, Panel, Card, BasicRow, Cover, Table } from 'coupon-components';
+import { Typography, Icon, Panel, Card, BasicRow, Table } from 'coupon-components';
+import PanelCampaign from 'Components/PanelCampaign/PanelCampaign';
 import { maxnum } from 'Utils/filters';
 import moment from 'moment';
 
@@ -152,60 +153,9 @@ class ShowCampaing extends Component {
           if (loading) return "Loading...";
           if (error) return notFound;
           const {campaign} = data;
-          const placeholderImage = 'https://www.ocf.berkeley.edu/~sather/wp-content/uploads/2018/01/food--1200x600.jpg';
-          const imageCover = (campaign && campaign.image) || placeholderImage
-          const stylesStatus = campaign && campaign.status === "expired" ?
-                            {color: palette.dark, backgroundColor: palette.baseGrayMedium} :
-                            {color: palette.whiteColor, backgroundColor: palette.primaryColor};
           return (
             <div className={styles.view}>
-              <div className={styles.header}>
-                <Cover
-                  status={campaign && campaign.status}
-                  image={imageCover}
-                  leftLabel={intl.formatMessage({id: 'campaigns.show.labels.office'})}
-                  leftText= {campaign.office.address}
-                  rightLabel={intl.formatMessage({id: 'campaigns.show.labels.available'})}
-                  rightText={maxnum(campaign.totalCoupons)}
-                  stylesStatus={stylesStatus}
-                />
-              </div>
-              <div className={styles.panelInformation}>
-                <div className={styles.left}>
-                  <Typography.Label small bold>
-                    {intl.formatMessage({id: 'campaigns.show.labels.title'})}
-                  </Typography.Label>
-                  <Typography.Text>
-                    {campaign && campaign.title}
-                  </Typography.Text>
-                  <Typography.Label small bold>
-                    {intl.formatMessage({id: 'campaigns.show.labels.description'})}
-                  </Typography.Label>
-                  <Typography.Text small>
-                    {campaign && campaign.description}
-                  </Typography.Text>
-                </div>
-                <div className={styles.right}>
-                  <Typography.Label small bold>
-                    {intl.formatMessage({id: 'campaigns.show.labels.segmentation'})}
-                  </Typography.Label>
-                  <div className={styles.icons}>
-                    <Icon
-                      name="FaMale"
-                      color={palette.baseGray}
-                      size={25}
-                    />
-                    <Icon
-                      name="FaFemale"
-                      color={palette.baseGray}
-                      size={25}
-                    />
-                  </div>
-                  <Typography.Text small>
-                    {campaign && campaign.initialAgeRange} - {campaign && campaign.finalAgeRange} años
-                  </Typography.Text>
-                </div>
-              </div>
+              <PanelCampaign campaign={campaign}/>
               <Panel
                 title={intl.formatMessage({id: 'campaigns.show.panel.title'})}
                 className={styles.hunters}>
